refactor: add tablegroup package
